Big Cleanup. READ THE DESCRIPTION.
- reverted back a lot of changes in order to get project working..
PLEASE READ BELOW
- cleaned up the project hierarchy - i hope it isn't too hard to find
what you have been working on. The .unity files are now in the /Scenes/
directory. If there are connections that are not correct i'm sorry cause
the whole project was screwed for me so i had to guess alot of stuff.
- NOTE: if you have made changes to anything PLEASE don't try and commit
your changes! All of the files are in totally different locations so you
will have to merge your changes manually after this change!
- Made the project use META files so that this doesn't happen again.
- Apart from that reverted my old change and put it back in. So
carriages can attach and detach.
- Commented out kit's test damage code on the carriages.